Update mod_clamav.c
The current implementation of ClamAV only works for chroot, but not for "virtual chroot" done by mod_vroot, which at this point in the upload-process does not set the "session.chroot_path". Thus "abs_path" would always contain a relative path if mod_vroot is uses and the scan would fail.
My proposed patch fixes this issue by determining the real path of the uploaded file, using the file descriptor before the file is closed.